Initial Draft of Claimant Response

+++ src/main/features/claimant-response/routes/check-and-send.ts
@@ -17,6 +17,11 @@ import { YesNoOption } from 'claims/models/response/core/yesNoOption'
 import { PaymentIntention } from 'claims/models/response/core/paymentIntention'
 import { MediationDraft } from 'mediation/draft/mediationDraft'
 import { FeatureToggles } from 'utils/featureToggles'
+import { SignatureType } from 'common/signatureType'
+import { Form } from 'forms/form'
+import { StatementOfTruth } from 'claimant-response/form/models/statementOfTruth'
+import { FormValidator } from 'forms/validation/formValidator'
+import { ResponseType } from 'claims/models/response/responseType'
 
 function getPaymentIntention (draft: DraftClaimantResponse, claim: Claim): PaymentIntention {
   const response: FullAdmissionResponse | PartialAdmissionResponse = claim.response as FullAdmissionResponse | PartialAdmissionResponse
@@ -44,37 +49,69 @@ function getPaymentIntention (draft: DraftClaimantResponse, claim: Claim): Payme
   return response.paymentIntention
 }
 
+function deserializerFunction (value: any): StatementOfTruth {
+  return StatementOfTruth.fromObject(value)
+}
+
+function renderView (form: Form<StatementOfTruth>, res: express.Response): void {
+  const draft: Draft<DraftClaimantResponse> = res.locals.claimantResponseDraft
+  const mediationDraft: Draft<MediationDraft> = res.locals.mediationDraft
+  const directionsQuestionnaireDraft = res.locals.directionsQuestionnaireDraft
+  const claim: Claim = res.locals.claim
+  const alreadyPaid: boolean = StatesPaidHelper.isResponseAlreadyPaid(claim)
+  const paymentIntention: PaymentIntention = alreadyPaid ? undefined : getPaymentIntention(draft.document, claim)
+  const dqsEnabled: boolean = (FeatureToggles.isEnabled('directionsQuestionnaire') && ((claim.response.responseType === ResponseType.PART_ADMISSION)
+    || (claim.response.responseType === ResponseType.FULL_DEFENCE)))
+  let datesUnavailable: string[]
+  if (dqsEnabled) {
+    datesUnavailable = directionsQuestionnaireDraft.document.availability.unavailableDates.map(date => date.toMoment().format('LL'))
+  }
+
+  const statementOfTruthType = dqsEnabled ? SignatureType.DIRECTION_QUESTIONNAIRE : SignatureType.RESPONSE
+  form.model.type = dqsEnabled ? SignatureType.DIRECTION_QUESTIONNAIRE : form.model.type
+
+  res.render(Paths.checkAndSendPage.associatedView, {
+    draft: draft.document,
+    claim: claim,
+    form: form,
+    totalAmount: AmountHelper.calculateTotalAmount(claim, res.locals.draft.document),
+    paymentIntention: paymentIntention,
+    alreadyPaid: alreadyPaid,
+    amount: alreadyPaid ? StatesPaidHelper.getAlreadyPaidAmount(claim) : undefined,
+    mediationEnabled: FeatureToggles.isEnabled('mediation'),
+    directionsQuestionnaireEnabled: dqsEnabled,
+    mediationDraft: mediationDraft.document,
+    directionsQuestionnaireDraft: directionsQuestionnaireDraft.document,
+    datesUnavailable: datesUnavailable,
+    statementOfTruthType: statementOfTruthType
+  })
+
+}
+
 /* tslint:disable:no-default-export */
 export default express.Router()
   .get(
     Paths.checkAndSendPage.uri,
     AllClaimantResponseTasksCompletedGuard.requestHandler,
     ErrorHandling.apply(async (req: express.Request, res: express.Response) => {
-      const draft: Draft<DraftClaimantResponse> = res.locals.claimantResponseDraft
-      const mediationDraft: Draft<MediationDraft> = res.locals.mediationDraft
-      const claim: Claim = res.locals.claim
-      const alreadyPaid: boolean = StatesPaidHelper.isResponseAlreadyPaid(claim)
-      const paymentIntention: PaymentIntention = alreadyPaid ? undefined : getPaymentIntention(draft.document, claim)
-      res.render(Paths.checkAndSendPage.associatedView, {
-        draft: draft.document,
-        claim: claim,
-        totalAmount: AmountHelper.calculateTotalAmount(claim, res.locals.draft.document),
-        paymentIntention: paymentIntention,
-        alreadyPaid: alreadyPaid,
-        amount: alreadyPaid ? StatesPaidHelper.getAlreadyPaidAmount(claim) : undefined,
-        mediationEnabled: FeatureToggles.isEnabled('mediation'),
-        mediationDraft: mediationDraft.document
-      })
+      const form = new Form(new StatementOfTruth())
+      renderView(form, res)
     })
   )
   .post(
     Paths.checkAndSendPage.uri,
     AllClaimantResponseTasksCompletedGuard.requestHandler,
+    FormValidator.requestHandler(undefined, deserializerFunction),
     ErrorHandling.apply(async (req: express.Request, res: express.Response) => {
-      const claim: Claim = res.locals.claim
-      const draft: Draft<DraftClaimantResponse> = res.locals.claimantResponseDraft
-      const user: User = res.locals.user
-      await new ClaimStoreClient().saveClaimantResponse(claim, draft, user)
-      await new DraftService().delete(draft.id, user.bearerToken)
-      res.redirect(Paths.confirmationPage.evaluateUri({ externalId: claim.externalId }))
+      const form: Form<StatementOfTruth> = req.body
+      if (form.hasErrors()) {
+        renderView(form, res)
+      } else {
+        const claim: Claim = res.locals.claim
+        const draft: Draft<DraftClaimantResponse> = res.locals.claimantResponseDraft
+        const user: User = res.locals.user
+        await new ClaimStoreClient().saveClaimantResponse(claim, draft, user)
+        await new DraftService().delete(draft.id, user.bearerToken)
+        res.redirect(Paths.confirmationPage.evaluateUri({ externalId: claim.externalId }))
+      }
     }))
